+++ src/mpi/send_recv.hpp
@@ -45,19 +45,13 @@ void Send_core(monolish::vector<T> &vec, int dst, int tag, MPI_Comm comm,
 
 #if defined MONOLISH_USE_GPU
   if ((gpu_sync == true) && (vec.get_device_mem_stat() == true)) {
-    vec.recv();
+    vec.nonfree_recv();
   }
 #endif
 
 #if defined MONOLISH_USE_MPI
   MPI_Send(vec.data(), vec.size(), internal::mpi::get_type(vec[0]), dst, tag,
            comm);
-#endif
-
-#if defined MONOLISH_USE_GPU
-  if (gpu_sync == true) {
-    vec.send();
-  }
 #endif
 
   logger.util_out();
@@ -109,12 +103,6 @@ MPI_Status Recv_core(monolish::vector<T> &vec, int src, int tag, MPI_Comm comm,
                      bool gpu_sync) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
-
-#if defined MONOLISH_USE_GPU
-  if ((gpu_sync == true) && (vec.get_device_mem_stat() == true)) {
-    vec.recv();
-  }
-#endif
 
 #if defined MONOLISH_USE_MPI
   MPI_Status stat;